Fix issue parsing format string w/ group separator
Fixes #656

Fix issue parsing log message format strings with placeholders containing a format string with a comma ",", such as "{setupTime:#,#}" which uses a custom numeric format string with a group separator. This fix uses the index of the first first comma "," or colon ":" found as the delimiter instead of ignoring the colon ":" if a comma "," was found.